% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{A different as.numeric}
\usage{
as_numeric(x, only.on = NULL, coerce.warning = TRUE)
}
\arguments{
\item{x}{An object to coerce to type numeric.}

\item{only.on}{Selects a subset of the data to which to apply as_numeric.}

\item{coerce.warning}{Whether to warn the user if coercion happened, defaults to TRUE}
}
\value{
An object of same structure as x, with elements coerced to numeric.
}
\description{
A function to coerce everything to type numeric, without giving an NA result.
}
\note{
Works with character vectors like "0,5145" without giving a warning or returning an NA.
}
\examples{

 # observe what as.numeric does
 A <- c("2,4","0.36","0.45","9,7")
 as.numeric(A)
 # and what as_numeric does
 as_numeric(A)
 # further if A is a factor
 A <- as.factor(c(10,2.7,5.5))
 as.numeric(A) # this is basically terrible and not what you want (most likely)
 as_numeric(A) # this works
 # further this function works happily with lists and matrices (and everything else)
 # and it will throw a warning and omit where the coercion would produce an NA


}
