% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_variable}
\alias{check_variable}
\title{Helper function which checks variable compatibility with provided type description}
\usage{
check_variable(var_obj, var_type, var_name = "foo")
}
\arguments{
\item{var_obj}{Variable to be checked}

\item{var_type}{Variable type description to be matched}

\item{var_name}{Variable name for logging purposes}
}
\value{
TRUE if types checks passed, FALSE otherwise. Also function writes detailed output to console.
}
\description{
Helper function which checks variable compatibility with provided type description
}
