% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typer.R
\name{compileFunction}
\alias{compileFunction}
\title{Remove the type decorators of a function and add the type checks in the body}
\usage{
compileFunction(fun, name = "[function]")
}
\arguments{
\item{fun}{[\code{function}]\cr
The function to compile.}

\item{name}{[\code{character(1)}]\cr
Name of the function to use in error messages.}
}
\value{
[\code{function}] The function with removed type decorators and added
type checks.
}
\description{
See \code{\link[=compileTypes]{compileTypes()}} for details on decorators. This is mostly used
internally, but can also be used to explicitly implement type checking for
a given function.
}
