% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helper.R
\name{.simulate_typing}
\alias{.simulate_typing}
\title{Internal Helper: Type Out Lines}
\usage{
.simulate_typing(lines, delay = 0.05)
}
\arguments{
\item{lines}{A character vector of script lines to simulate typing.}

\item{delay}{The delay (in seconds) between typing each character (default: 0.05).}
}
\description{
An internal function that simulates typing out a vector of strings.
}
\keyword{internal}
