% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{generate_func_doc}
\alias{generate_func_doc}
\title{Generate function HTML doc basing on types description}
\usage{
generate_func_doc(func_body, func_name)
}
\arguments{
\item{func_body}{Function body - decorator function which is returned by \code{\link{check_function}}}

\item{func_name}{Function name}
}
\value{
Character vector with HTML text which contains function doc section
}
\description{
Helper function which generates HTML document section for function input / output parameters
}
