% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{generate_package_doc}
\alias{generate_package_doc}
\title{Generate package HTML docs basing on types description}
\usage{
generate_package_doc(package_name)
}
\arguments{
\item{package_name}{Package name}
}
\value{
Character vector with HTML text which contains package functions documentation
}
\description{
Helper function which generates HMTL documentation for published package functions with input / output checkers
}
\examples{
\dontrun{
  html_report <- generate_package_doc("demopack")
  html_file <- tempfile(fileext = ".html")
  writeChar(html_report, html_file)
  browseURL(html_file)
}

}
