% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_functions.R
\name{generate_variable_doc}
\alias{generate_variable_doc}
\title{Generate variable HTML doc basing on types description}
\usage{
generate_variable_doc(var_type, var_name)
}
\arguments{
\item{var_type}{List with variable type attributes}

\item{var_name}{Variable name}
}
\value{
Character vector with HTML text which contains variable doc section
}
\description{
Helper function which generates HTML document section for provided variable type description
}
