% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_type_errors}
\alias{get_type_errors}
\title{Check a file for type errors}
\usage{
get_type_errors(filename, full_results = FALSE)
}
\arguments{
\item{filename}{File to check}

\item{full_results}{if TRUE, return all results. If FALSE, only return errors.}
}
\value{
A list containing the errors
}
\description{
Check for inconsistent use of types within one file.
}
\details{
This function checks for inconsistent uses
of types within a file. It checks all function calls in
the file to see whether arguments in the form
of literals, parameters, or return values of other calls
match the type the argument is documented to be.

The function returns a list of results and warns when errors
are encountered.
}
