% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typer.R
\name{\%t\%}
\alias{\%t\%}
\title{A type coercion operator}
\usage{
obj \%t\% new_type
}
\arguments{
\item{obj}{The object to coerce into a new type.}

\item{new_type}{The character string (or string shorthand) giving the new type required.}
}
\value{
The coerced object.
}
\description{
Allows the quick changing of types, using the new type coercions introduced in the package.
}
\details{
Behaves a bit more smartly than base coercion (f.e. numbers containing a "," as a separator get coerced to type numeric)
}
\note{
Accepts both full argument names: "numeric","character","logical",
                                        "integer","double","complex",
                                        "list","unlist","NULL","environment",
                                        "symbol","expression","data.frame",
                                        "vector","factor", "data.table"

                 as well as shorthands: "num","char","logi",
                                        "int","dbl","cmplx",
                                        "lst","unlst","NUL","env",
                                        "sym","expr","dfm",
                                        "vec","fac","dtb"
}
\examples{


A <- c( "0,7",0.9,"15","1235","0.66","plane" )

# there are shorthands
A \%t\% "num"
A \%t\% "character"
A \%t\% "factor"

# NOTE that you can create your own shorthands using global environment objects
# for interactive use if you want to
# (Please never do this for anything that is not throwaway code, meaning code that
# gets reused or run multiple times;
# it WILL cause confusion and IT WILL make people angry. )
n <- "numeric"
A \%t\% n


}
