% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_builders.R
\name{type.complex}
\alias{type.complex}
\title{Type Builder for Complex Numbers Vector}
\usage{
type.complex(len = NA_integer_, allow.na = NA, desc = NULL)
}
\arguments{
\item{len}{Expected length of vector to be checked. Check will be skipped if len is NA}

\item{allow.na}{Logical flag which indicates whether vector may contain NA values}

\item{desc}{Optional string with variable purpose descritption}
}
\description{
Type Builder for Complex Numbers Vector
}
