% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_builders.R
\name{type.data.frame}
\alias{type.data.frame}
\title{Type Builder for Data Frame}
\usage{
type.data.frame(
  cols,
  nrow = NA_integer_,
  ncol = NA_integer_,
  key = NA_character_,
  desc = NULL
)
}
\arguments{
\item{cols}{Named list with data frame columns descriptions}

\item{nrow}{Expected count of rows in data frame}

\item{ncol}{Expected count of columns in data frame}

\item{key}{Character vector with data frame key fields, may be skipped}

\item{desc}{Optional string with variable purpose descritption}

\item{allow.na}{Logical flag which indicates whether vector may contain NA values}
}
\description{
Type Builder for Data Frame
}
