% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_builders.R
\name{type.data.table}
\alias{type.data.table}
\title{Type Builder for Data Table}
\usage{
type.data.table(
  cols,
  nrow = NA_integer_,
  ncol = NA_integer_,
  key = NA_character_,
  desc = NULL
)
}
\arguments{
\item{cols}{Named list with data table columns descriptions}

\item{nrow}{Expected count of rows in data table}

\item{ncol}{Expected count of columns in data table}

\item{key}{Character vector with data table key fields, may be skipped}

\item{desc}{Optional string with variable purpose descritption}

\item{allow.na}{Logical flag which indicates whether vector may contain NA values}
}
\description{
Type Builder for Data Table
}
