% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_builders.R
\name{type.expr}
\alias{type.expr}
\title{Type Builder for Raw Vector}
\usage{
type.expr(desc = NULL)
}
\arguments{
\item{desc}{Optional string with variable purpose descritption}

\item{len}{Expected length of vector to be checked. Check will be skipped if len is NA}

\item{allow.na}{Logical flag which indicates whether vector may contain NA values}
}
\description{
Type Builder for Raw Vector
}
