% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_builders.R
\name{type.factor}
\alias{type.factor}
\title{Type Builder for Vector of Factors}
\usage{
type.factor(levels = NULL, desc = NULL)
}
\description{
This wraps an arbitrary function so that its input and output could be validate against predefined type safety rules. If either fails to pass, error is thrown.
}
