% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_builders.R
\name{type.integer}
\alias{type.integer}
\title{Type Builder for Integer Vector}
\usage{
type.integer(len = NA_integer_, allow.na = TRUE, desc = NULL)
}
\arguments{
\item{len}{Expected length of vector to be checked. Check will be skipped if len is NA}

\item{allow.na}{Logical flag which indicates whether vector may contain NA values}

\item{desc}{Optional string with variable purpose descritption}
}
\description{
Type Builder for Integer Vector
}
