% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_builders.R
\name{type.matrix.character}
\alias{type.matrix.character}
\title{Type Builder for Character Matrix}
\usage{
type.matrix.character(
  nrow = NA_integer_,
  ncol = NA_integer_,
  allow.na = TRUE,
  desc = NULL
)
}
\arguments{
\item{allow.na}{Logical flag which indicates whether vector may contain NA values}

\item{desc}{Optional string with variable purpose descritption}

\item{len}{Expected length of vector to be checked. Check will be skipped if len is NA}
}
\description{
Type Builder for Character Matrix
}
