% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeR-package.R
\docType{package}
\name{typeR-package}
\alias{typeR-package}
\title{typeR: Simulate Typing Script}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simulates typing of R script files for presentations and demonstrations. Provides character-by-character animation with optional live code execution. Supports R scripts (.R), R Markdown (.Rmd), and Quarto (.qmd) documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://Fgazzelloni.github.io/typeR/}
  \item \url{https://github.com/Fgazzelloni/typeR}
  \item Report bugs at \url{https://github.com/Fgazzelloni/typeR/issues}
}

}
\author{
\strong{Maintainer}: Federica Gazzelloni \email{fede.gazzelloni@gmail.com}

}
\keyword{internal}
