% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{A different as.numeric}
\usage{
as_numeric(x, only.on = NULL, coerce.warning = TRUE)
}
\arguments{
\item{x}{An object to coerce to type numeric.}

\item{only.on}{Selects a subset of the data to which to apply as_numeric.}

\item{coerce.warning}{Whether to warn the user if coercion happened, defaults to TRUE}
}
\value{
An object of same structure as x, with elements coerced to numeric.
}
\description{
A function to coerce everything to type numeric, without giving an NA result.
}
\note{
Works with character vectors like "0,5145" without giving a warning or returning an NA.
}
