% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_type_errors}
\alias{get_type_errors}
\title{Check a file for type errors}
\usage{
get_type_errors(filename)
}
\arguments{
\item{filename}{File to check}
}
\value{
A list containing the errors
}
\description{
Check for inconsistent use of types within one file.
}
\details{
This function checks for inconsistent uses
of types within a file. It checks all function calls in
the file to see whether arguments in the form
of literals, parameters, or return values of other calls
match the type the argument is documented to be.

The return value is a list of errors, with each error reported
in the following format: `error$fn`, the name of the function that was
given the wrong argument; `error$call`: full call object that produced
the error; `error$type`, which is "literal", "parameter", or "returnval";
`error$expected`, the expected type, as given by `typeof()`; and
`error$got`, the received type, as given by `typeof()`.
}
