% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeR-package.R
\docType{package}
\name{typeR-package}
\alias{typeR-package}
\title{typeR: Simulate Typing Script}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package to simulate typing of R script files for presentations and demonstrations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fgazzelloni.github.io/typeR}
  \item \url{https://github.com/fgazzelloni/typeR}
  \item Report bugs at \url{https://github.com/fgazzelloni/typeR/issues}
}

}
\author{
\strong{Maintainer}: Federica Gazzelloni \email{fede.gazzelloni@gmail.com}

}
\keyword{internal}
