% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeR.R
\name{typeR}
\alias{typeR}
\title{Simulate Typing of R Scripts}
\usage{
typeR(file, delay = 0.05)
}
\arguments{
\item{file}{Path to the R script file to simulate typing.}

\item{delay}{The delay (in seconds) between typing each character. Default is 0.05 seconds.}
}
\description{
Simulates typing out the contents of an R script file, character by character,
to create a "typing" animation effect.
}
\examples{
# Simulate typing of a script
# typeR("example_script.R", delay = 0.1)
}
