% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeR.R
\name{typeR}
\alias{typeR}
\title{Simulate Typing of an R Script File}
\usage{
typeR(file, delay = 0.05)
}
\arguments{
\item{file}{Path to the R script file to simulate typing.}

\item{delay}{The delay (in seconds) between typing each character (default: 0.05).}
}
\description{
Simulates typing out the content of an R script file, line by line,
character by character, to create an animation effect for live coding
presentations or educational demonstrations.
}
\examples{
# Simulate typing of a script
# typeR("example_script.R", delay = 0.1)
}
